requirements.txt
scikit-learn
opencv-python-headless
streamlit
numpy
pillow
io